requirements.txt
Flask
openai
python-dotenv
mysql-connector-python